Updated Readme.md
I tried having an options object as shown in your readme without a method property (only with host, path, and query). It threw a TypeError that it cannot read property 'toUpperCase' of undefined which is from line 148 of your canonicalString method in /lib/aws4.js. 

After I prepped the opt object with a method, it is now upcased everything properly. I think this will make it more painless for people to use your module. Thanks for writing it! Credentials are a pain in the butt!